+++ docdoku-plm-conversion-service/conversion-service/src/main/java/com/docdoku/plm/conversion/service/App.java
@@ -201,7 +201,7 @@ public class App {
             partApi.sendConversionResult(workspaceId, partNumber, partVersion, conversionResult);
             LOGGER.info("Conversion callback success");
         } catch (ApiException e) {
-            LOGGER.severe("Conversion callback error");
+            LOGGER.severe("Conversion callback error: " + e.getCode() + " " + e.getMessage() + " " + e.getResponseBody());
         }
     }